Fix early processing integration events at startup
