+++ src/BuildingBlocks/EventBus/EventBus/Abstractions/IEventBus.cs
@@ -20,5 +20,7 @@ namespace Microsoft.eShopOnContainers.BuildingBlocks.EventBus.Abstractions
         void Unsubscribe<T, TH>()
             where TH : IIntegrationEventHandler<T>
             where T : IntegrationEvent;
+
+        void Start();
     }
 }
