+++ src/BuildingBlocks/EventBus/EventBusDependencyInjection/Extensions/EventBusServiceBusServiceCollectionExtensions.cs
@@ -1,0 +1,59 @@
+﻿
+namespace Microsoft.Extensions.DependencyInjection
+{
+    using Autofac;
+    using Microsoft.Azure.ServiceBus;
+    using Microsoft.eShopOnContainers.BuildingBlocks.EventBus;
+    using Microsoft.eShopOnContainers.BuildingBlocks.EventBus.Abstractions;
+    using Microsoft.eShopOnContainers.BuildingBlocks.EventBusServiceBus;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Logging;
+    using System.Linq;
+    using System.Reflection;
+    public static class EventBusServiceBusServiceCollectionExtensions
+    {
+        public static IServiceCollection AddEventBusServiceBus(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddServiceBus(configuration);
+
+            var subscriptionClientName = configuration["SubscriptionClientName"];
+
+            services.AddSingleton<IEventBus, EventBusServiceBus>(sp =>
+            {
+                var serviceBusPersisterConnection = sp.GetRequiredService<IServiceBusPersisterConnection>();
+                var iLifetimeScope = sp.GetRequiredService<ILifetimeScope>();
+                var logger = sp.GetRequiredService<ILogger<EventBusServiceBus>>();
+                var eventBusSubcriptionsManager = sp.GetRequiredService<IEventBusSubscriptionsManager>();
+
+                return new EventBusServiceBus(serviceBusPersisterConnection, logger,
+                    eventBusSubcriptionsManager, subscriptionClientName, iLifetimeScope);
+            });
+
+
+            services.AddSingleton<IEventBusSubscriptionsManager, InMemoryEventBusSubscriptionsManager>();
+
+            return services;
+        }
+
+        private static IServiceCollection AddServiceBus(this IServiceCollection services, IConfiguration Configuration)
+        {
+            services.AddSingleton<IServiceBusPersisterConnection>(sp =>
+            {
+                var logger = sp.GetRequiredService<ILogger<DefaultServiceBusPersisterConnection>>();
+
+                var serviceBusConnectionString = Configuration["EventBusConnection"];
+                var serviceBusConnection = new ServiceBusConnectionStringBuilder(serviceBusConnectionString);
+
+                return new DefaultServiceBusPersisterConnection(serviceBusConnection, logger);
+            });
+
+            return services;
+        }
+
+
+
+
+    }
+}
+
