+++ src/BuildingBlocks/EventBus/EventBusDependencyInjection/Extensions/EventBusServiceCollectionExtensions.cs
@@ -1,0 +1,38 @@
+﻿
+namespace Microsoft.Extensions.DependencyInjection
+{
+    using Autofac;
+    using Microsoft.eShopOnContainers.BuildingBlocks.EventBus;
+    using Microsoft.eShopOnContainers.BuildingBlocks.EventBus.Abstractions;
+    using Microsoft.eShopOnContainers.BuildingBlocks.EventBusRabbitMQ;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Logging;
+    using RabbitMQ.Client;
+    using System.Linq;
+    using System.Reflection;
+
+    public static class EventBusServiceCollectionExtensions
+    {
+        public static IServiceCollection AddEventBus(this IServiceCollection services, IConfiguration configuration)
+        {
+            if (configuration.GetValue<bool>("AzureServiceBusEnabled"))
+            {
+                services.AddEventBusServiceBus(configuration);
+            }
+            else
+            {
+                services.AddEventBusRabbitMq(configuration);
+            }
+
+           return services;
+        }
+
+       
+
+
+
+
+    }
+}
+
