+++ src/eShopOnContainers-ServicesAndWebApps.sln
@@ -157,6 +157,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "config", "config", "{3ABEEE
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ordering.BackgroundTasks", "Services\Ordering\Ordering.BackgroundTasks\Ordering.BackgroundTasks.csproj", "{D4DBA4A3-E4A5-4D9D-8ACF-F38F7D506191}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EventBusDependencyInjection", "BuildingBlocks\EventBus\EventBusDependencyInjection\EventBusDependencyInjection.csproj", "{3B82FD32-03B2-48D5-856D-17664C51F6A7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Ad-Hoc|Any CPU = Ad-Hoc|Any CPU
@@ -1707,6 +1709,54 @@ Global
 		{D4DBA4A3-E4A5-4D9D-8ACF-F38F7D506191}.Release|x64.Build.0 = Release|Any CPU
 		{D4DBA4A3-E4A5-4D9D-8ACF-F38F7D506191}.Release|x86.ActiveCfg = Release|Any CPU
 		{D4DBA4A3-E4A5-4D9D-8ACF-F38F7D506191}.Release|x86.Build.0 = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|Any CPU.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|ARM.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|iPhone.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|iPhone.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|x64.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|x64.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|x86.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Ad-Hoc|x86.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|Any CPU.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|ARM.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|ARM.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|iPhone.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|iPhone.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|x64.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|x64.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|x86.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.AppStore|x86.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|ARM.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|x64.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Debug|x86.Build.0 = Debug|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|ARM.ActiveCfg = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|ARM.Build.0 = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|iPhone.Build.0 = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|x64.ActiveCfg = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|x64.Build.0 = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|x86.ActiveCfg = Release|Any CPU
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1772,6 +1822,7 @@ Global
 		{882A8F3A-C61F-4C44-86DD-A5A258714BF2} = {77849D35-37D4-4802-81DC-9477B2775A40}
 		{3ABEEE8C-35E0-4185-9825-C44326151F5B} = {882A8F3A-C61F-4C44-86DD-A5A258714BF2}
 		{D4DBA4A3-E4A5-4D9D-8ACF-F38F7D506191} = {0BD0DB92-2D98-44D9-9AC0-C59186D59B0B}
+		{3B82FD32-03B2-48D5-856D-17664C51F6A7} = {807BB76E-B2BB-47A2-A57B-3D1B20FF5E7F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {25728519-5F0F-4973-8A64-0A81EB4EA8D9}
